# doc/api-documentation.yml
---
openapi: 3.0.0
info:
  version: 2.2.0
  title: Hacking Health
  description: ''
servers:
- url: https://api.hcgateway.shuchir.dev
paths:
  "/login":
    post:
      tags:
        - Authentication
      operationId: get_token
      description: Login a user and get a token to use for authentiction.
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/loginParameters"
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/loginObject"
  "/refresh":
    post:
      tags:
        - Authentication
      operationId: refresh_token
      description: Refresh a user token.
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              properties:
                refresh:
                  nullable: false
                  type: string
                  description: The refresh token you got when you logged in.
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/loginObject"
  "/revoke":
    delete:
      tags:
        - Authentication
      operationId: revoke_token
      description: Revoke all user tokens and refresh tokens immediately. Pass a currently valid token through the Authentication header to revoke your user.
      parameters: []
      security:
      - bearerAuth: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/silentObject"
  "/sync/{method}":
    post:
      tags:
        - Sync
      operationId: sync
      description: >
        Sync data to the database. Accepts both individual Health Connect records 
        and aggregated payloads (recommended for better performance).
        
        For high-frequency data types (steps, heartRate, distance, etc.), consider 
        using aggregated payloads to reduce network traffic and improve sync speed.
        
        The backend validates that the record type is in the canonical allow-list 
        and accepts both camelCase (canonical) and PascalCase (Health Connect) formats.
      security:
      - bearerAuth: []
      parameters:
      - in: path
        name: method
        required: true
        schema:
          type: string
        description: 'Can be one of the following canonical record types (case-insensitive): 
          activeCaloriesBurned, basalBodyTemperature, basalMetabolicRate, bloodGlucose, 
          bloodPressure, bodyFat, bodyTemperature, boneMass, cervicalMucus, distance, 
          exerciseSession, elevationGained, floorsClimbed, heartRate, height, hydration, 
          leanBodyMass, menstruationFlow, menstruationPeriod, nutrition, ovulationTest, 
          oxygenSaturation, power, respiratoryRate, restingHeartRate, sleepSession, speed, 
          steps, stepsCadence, totalCaloriesBurned, vo2Max, weight, wheelchairPushes'
      requestBody:
        content:
          application/json:
            schema:
              properties:
                data:
                  nullable: false
                  description: >
                    The data to sync to the database. Can be:
                    1. Array of Health Connect record objects (traditional format)
                    2. Array of aggregated payloads (recommended - see aggregatedPayload schema)
                    3. Single record object
                    4. Single aggregated payload object
                  oneOf:
                    - type: array
                      items:
                        type: object
                    - type: object
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/silentObject"
    delete:
      tags:
        - Sync
      operationId: deleteFromDb
      description: Delete records from the Hacking Health database.
      security:
      - bearerAuth: []
      parameters:
      - in: path
        name: method
        required: true
        schema:
          type: string
        description: 'Can be one of the following: activeCaloriesBurned, basalBodyTemperature,
          basalMetabolicRate, bloodGlucose, bloodPressure, bodyFat, bodyTemperature,
          boneMass, cervicalMucus, distance, exerciseSession, elevationGained, floorsClimbed,
          heartRate, height, hydration, leanBodyMass, menstruationFlow, menstruationPeriod,
          nutrition, ovulationTest, oxygenSaturation, power, respiratoryRate, restingHeartRate,
          sleepSession, speed, steps, stepsCadence, totalCaloriesBurned, vo2Max, weight,
          wheelchairPushes'
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/deleteParameters"
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/silentObject"
  "/sync/{method}/fetch":
    post:
      tags:
        - Sync
      operationId: fetch
      description: Get data from the database.
      security:
      - bearerAuth: []
      parameters:
      - in: path
        name: method
        required: true
        schema:
          type: string
        description: 'Can be one of the following: activeCaloriesBurned, basalBodyTemperature,
          basalMetabolicRate, bloodGlucose, bloodPressure, bodyFat, bodyTemperature,
          boneMass, cervicalMucus, distance, exerciseSession, elevationGained, floorsClimbed,
          heartRate, height, hydration, leanBodyMass, menstruationFlow, menstruationPeriod,
          nutrition, ovulationTest, oxygenSaturation, power, respiratoryRate, restingHeartRate,
          sleepSession, speed, steps, stepsCadence, totalCaloriesBurned, vo2Max, weight,
          wheelchairPushes'
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/fetchParameters"
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/dataObject"
  "/sync/{method}/push":
    put:
      tags:
        - Sync
      operationId: pushData
      description: Insert data into your local Health Connect store. This method does
        not add the data to the Hacking Health database. This method is only used to push
        data to your local Health Connect store. This is so you can sync + fetch to
        verify the record has been added.
      security:
      - bearerAuth: []
      parameters:
      - in: path
        name: method
        required: true
        schema:
          type: string
        description: 'Can be one of the following: activeCaloriesBurned, basalBodyTemperature,
          basalMetabolicRate, bloodGlucose, bloodPressure, bodyFat, bodyTemperature,
          boneMass, cervicalMucus, distance, exerciseSession, elevationGained, floorsClimbed,
          heartRate, height, hydration, leanBodyMass, menstruationFlow, menstruationPeriod,
          nutrition, ovulationTest, oxygenSaturation, power, respiratoryRate, restingHeartRate,
          sleepSession, speed, steps, stepsCadence, totalCaloriesBurned, vo2Max, weight,
          wheelchairPushes'
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/pushParameters"
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/silentObject"
  "/sync/{method}/delete":
    delete:
      tags:
        - Sync
      operationId: deleteData
      description: Request deletion for a record. This method does not delete the
        record from the Hacking Health database. This method is only used to request deletion
        of a record from your local Health Connect store. The app will delete it from
        the database using the sync method once it has deleted the record(s) locally.
      security:
      - bearerAuth: []
      parameters:
      - in: path
        name: method
        required: true
        schema:
          type: string
        description: 'Can be one of the following: activeCaloriesBurned, basalBodyTemperature,
          basalMetabolicRate, bloodGlucose, bloodPressure, bodyFat, bodyTemperature,
          boneMass, cervicalMucus, distance, exerciseSession, elevationGained, floorsClimbed,
          heartRate, height, hydration, leanBodyMass, menstruationFlow, menstruationPeriod,
          nutrition, ovulationTest, oxygenSaturation, power, respiratoryRate, restingHeartRate,
          sleepSession, speed, steps, stepsCadence, totalCaloriesBurned, vo2Max, weight,
          wheelchairPushes'
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/deleteParameters"
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/silentObject"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    dataObject:
      properties:
        _id:
          nullable: false
          type: string
          description: The id Health Connect stored the object with.
        data:
          nullable: false
          type: object
          description: The actual data stored in Health Connect
        id:
          nullable: false
          type: string
        start:
          nullable: false
          type: string
          format: date-time
          description: The start time of the data
        end:
          nullable: true
          type: string
          format: date-time
          description: The end time of the data. Only present if the data is a time-series.
        app:
          nullable: false
          type: string
          description: The app that created the data
    silentObject:
      properties:
        success:
          nullable: false
          type: boolean
          description: Sucessfully sent the data or not
        message:
          nullable: false
          type: string
          description: A confirmation message.
    loginObject:
      properties:
        token:
          nullable: false
          type: string
          description: The bearer token you should use for future requests.
        refresh:
          nullable: false
          type: string
          description: The refresh token you should use to get a new token after the current one expires.
        expiry:
          nullable: false
          type: string
          format: date-time
          description: The time the token expires. This is always 12 hours after the token was issued.
    loginParameters:
      properties:
        username:
          nullable: false
          type: string
          required: true
        password:
          nullable: false
          type: string
          required: true
    fetchParameters:
      properties:
        queries:
          nullable: true
          type: object
          description: Queries to use to filter the data. Must be a valid mongodb
            query. See https://www.mongodb.com/docs/languages/python/pymongo-driver/current/read/specify-a-query/
            for more information.
    pushParameters:
      properties:
        data:
          nullable: false
          type: array
          items:
            type: object
          description: The record to push. This can be an object or array of objects.
            See https://developer.android.com/reference/kotlin/androidx/health/connect/client/records/package-summary#classes
            for each data type's object structure.
    deleteParameters:
      properties:
        uuid:
          nullable: false
          type: array
          items:
            type: string
          description: The uuid of the record to delete. Can be a string or an array
            of strings.
    aggregatedPayload:
      description: >
        Aggregated health data payload for efficient bulk sync.
        Instead of sending per-minute records, data is aggregated into hourly or daily buckets.
        This reduces payload size and improves sync performance.
      properties:
        type:
          nullable: false
          type: string
          enum:
            - activeCaloriesBurned
            - basalBodyTemperature
            - basalMetabolicRate
            - bloodGlucose
            - bloodPressure
            - bodyFat
            - bodyTemperature
            - boneMass
            - cervicalMucus
            - distance
            - exerciseSession
            - elevationGained
            - floorsClimbed
            - heartRate
            - height
            - hydration
            - leanBodyMass
            - menstruationFlow
            - menstruationPeriod
            - nutrition
            - ovulationTest
            - oxygenSaturation
            - power
            - respiratoryRate
            - restingHeartRate
            - sleepSession
            - speed
            - steps
            - stepsCadence
            - totalCaloriesBurned
            - vo2Max
            - weight
            - wheelchairPushes
          description: The canonical record type (camelCase format)
        granularity:
          nullable: false
          type: string
          enum:
            - hour
            - day
          description: Time bucket granularity for aggregation
        periodStart:
          nullable: false
          type: string
          format: date-time
          description: Start of the aggregation period (ISO 8601)
        periodEnd:
          nullable: false
          type: string
          format: date-time
          description: End of the aggregation period (ISO 8601)
        data:
          nullable: false
          type: object
          description: >
            Aggregated data for the period. Structure varies by record type.
            Examples:
            - steps: { "totalSteps": 5430 }
            - heartRate: { "avgBpm": 72, "minBpm": 58, "maxBpm": 145, "samples": 120 }
            - sleepSession: { "totalMinutes": 458, "stagesMinutes": { "DEEP": 90, "LIGHT": 250, "REM": 118 } }
            - distance: { "totalMeters": 8420.5 }
            - activeCaloriesBurned: { "totalKcal": 523.4 }
        metadata:
          nullable: true
          type: object
          description: Optional metadata about the aggregated data (e.g., source app, device info)